@@ -1,13 +1,12 @@
 apply plugin: 'com.android.application'
 
 android {
-    compileSdkVersion 18
-    buildToolsVersion "29.0.2"
+    compileSdkVersion 29
 
     defaultConfig {
         applicationId "com.mobile.bummerzaehler"
-        minSdkVersion 11
-        targetSdkVersion 21
+        minSdkVersion 14
+        targetSdkVersion 29
     }
 
     buildTypes {
@@ -19,5 +18,5 @@ android {
 }
 
 dependencies {
-    implementation 'com.android.support:support-v4:18.0.0'
+    implementation 'androidx.legacy:legacy-support-v4:1.0.0'
 }
