@@ -2,9 +2,6 @@ package com.mobile.bummerzaehler;
 
 
 
-import java.util.ArrayList;
-import java.util.Random;
-
 import com.mobile.bummerzaehler.bummerl.AllBummerls2P;
 import com.mobile.bummerzaehler.bummerl.Bummerl2PController;
 import com.mobile.bummerzaehler.helper.HelperClass;
@@ -12,42 +9,22 @@ import com.mobile.bummerzaehler.oldgame.OldGame2P;
 import com.mobile.bummerzaehler.oldgame.OldGames2PController;
 import com.mobile.bummerzaehler.player.Player;
 
-import android.animation.ArgbEvaluator;
-import android.animation.ObjectAnimator;
 import android.animation.ValueAnimator;
 import android.animation.ValueAnimator.AnimatorUpdateListener;
-import android.app.Activity;
 import android.app.AlertDialog;
-import android.app.Notification;
-import android.app.NotificationManager;
-import android.app.PendingIntent;
-import android.app.TaskStackBuilder;
-import android.content.Context;
 import android.content.DialogInterface;
 import android.content.Intent;
-import android.content.SharedPreferences;
-import android.graphics.Color;
 import android.graphics.drawable.TransitionDrawable;
-import android.net.Uri;
-import android.opengl.Visibility;
 import android.os.Bundle;
 import android.preference.PreferenceManager;
-import android.support.v4.app.NotificationCompat;
-import android.view.KeyEvent;
-import android.view.Menu;
+import androidx.core.app.NotificationCompat;
+
 import android.view.MenuItem;
-import android.view.MotionEvent;
 import android.view.View;
 import android.view.View.OnClickListener;
 import android.view.WindowManager;
-import android.view.animation.Animation;
-import android.view.animation.Transformation;
-import android.widget.Button;
-import android.widget.EditText;
 import android.widget.FrameLayout;
-import android.widget.ImageView;
 import android.widget.LinearLayout;
-import android.widget.LinearLayout.LayoutParams;
 import android.widget.ScrollView;
 import android.widget.TextView;
 
